refactor: add more en&zh string values, start to write view for item list in home page
