+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.AddItemActivity" />
+        <activity android:name=".AddItemActivity" />
         <activity android:name=".AddDescription"
             android:windowSoftInputMode="adjustResize">
         </activity>
