+++ app/src/main/java/com/yuukidach/ucount/view/MainItemListView.java
@@ -1,0 +1,4 @@
+package com.yuukidach.ucount.view;
+
+public interface MainItemListView {
+}
