+++ app/src/main/res/values/strings.xml
@@ -3,6 +3,9 @@
     <string name="show_balance">Show Balance</string>
     <string name="monthly_income_title">Monthly Earn</string>
     <string name="monthly_expenditure_title">Monthly Cost</string>
+    <string name="delete_item_alarm">Are you sure to delete it?</string>
+    <string name="cancle">Cancle</string>
+    <string name="confirm">Confirm</string>
     <string name="add_cost_button_string">支出</string>
     <string name="add_earn_button_string">收入</string>
